Java and the front end full stack learning path

+++ gulimall-coupon/src/main/java/com/example/gulimall/coupon/entity/SeckillSkuRelationEntity.java
@@ -2,18 +2,14 @@ package com.example.gulimall.coupon.entity;
 
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-
-import java.math.BigDecimal;
-import java.io.Serializable;
-import java.util.Date;
 import lombok.Data;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 秒杀活动商品关联
  * 
- * @author jx512337
- * @email jx512337@gmail.com
- * @date 2020-08-03 00:45:31
  */
 @Data
 @TableName("sms_seckill_sku_relation")
@@ -44,11 +40,11 @@ public class SeckillSkuRelationEntity implements Serializable {
 	/**
 	 * 秒杀总量
 	 */
-	private BigDecimal seckillCount;
+	private Integer seckillCount;
 	/**
 	 * 每人限购数量
 	 */
-	private BigDecimal seckillLimit;
+	private Integer seckillLimit;
 	/**
 	 * 排序
 	 */
